Refactor generation into separate function
